+++ crates/miden-protocol/src/errors/mod.rs
@@ -762,6 +762,8 @@ impl PartialBlockchainError {
 pub enum TransactionScriptError {
     #[error("failed to assemble transaction script:\n{}", PrintDiagnostic::new(.0))]
     AssemblyError(Report),
+    #[error("failed to convert package to transaction script:\n{}", PrintDiagnostic::new(.0))]
+    PackageNotProgram(Report),
 }
 
 // TRANSACTION INPUT ERROR